Feat/manage user account

+++ src/main/java/com/backend/cms/controller/UserController.java
@@ -43,9 +43,10 @@ public class UserController {
         return UserDTO.fromUser(user);
     }
 
-    @RequestMapping(value = "{id}", method = RequestMethod.PUT)
+    @RequestMapping(value = "{id}", method = RequestMethod.PATCH)
     public UserDTO update(@PathVariable("id") String id, @Valid @RequestBody UpdateUserRequest request) {
         LOGGER.info("Updating user entry with information: {}", request);
+        authService.checkIfUserIsAdminOrThrowException();
         User user = userService.findUserFailIfNotFound(id);
         userService.updateUser(user, request);
         LOGGER.info("Updated user entry with information: {}", user);
@@ -55,6 +56,7 @@ public class UserController {
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
     public UserDTO delete(@PathVariable("id") String id) {
         LOGGER.info("Deleting user entry with id: {}", id);
+        authService.checkIfUserIsAdminOrThrowException();
         User user = userService.findUserFailIfNotFound(id);
         userRepository.delete(user);
         LOGGER.info("Deleted user entry with information: {}", user);
